+++ src/bika/lims/content/analysisrequest.py
@@ -167,14 +167,19 @@ schema = BikaSchema.copy() + Schema((
             ui_item="Title",
             catalog=CONTACT_CATALOG,
             # Base query - gets overridden with client-specific query at
-            #  runtime to include both client contacts and global contacts
+            # runtime to include both client contacts and global contacts
             query={
-                "getParentUID": "",
                 "is_active": True,
                 "sort_on": "sortable_title",
                 "sort_order": "ascending"
             },
             columns=[
+                {
+                    "name": "scope",
+                    "width": "10",
+                    "align": "center",
+                    "label": "",
+                },
                 {"name": "Title", "label": _("Name")},
                 {"name": "getEmailAddress", "label": _("Email")},
             ],
@@ -205,12 +210,17 @@ schema = BikaSchema.copy() + Schema((
             # Base query - gets overridden with client-specific query at
             # runtime to include both client contacts and global contacts
             query={
-                "getParentUID": "",
                 "is_active": True,
                 "sort_on": "sortable_title",
                 "sort_order": "ascending"
             },
             columns=[
+                {
+                    "name": "scope",
+                    "width": "10",
+                    "align": "center",
+                    "label": "",
+                },
                 {"name": "Title", "label": _("Name")},
                 {"name": "getEmailAddress", "label": _("Email")},
             ],